pyproject.toml
[project]
name = "clc3-rag-tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astra-haystack>=0.10.1",
    "fastapi[standard]>=0.115.8",
    "hayhooks>=0.0.19",
    "haystack-ai>=2.9",
    "ipykernel>=6.29.5",
    "ollama-haystack>=2.3.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-sdk>=1.29.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
    "traceloop-sdk>=0.37.1",
    "weaviate-haystack>=4.0.2",
]
